clarifying known issue for 1.3.1
changed "if you are iterating on the Async Queue, the cluster operator..." to "if you are iterating on the Async Queue, be aware that the cluster operator..." because we don't want to imply that the cluster operator's ability (or lack of it) to delete the queue depends upon whether the user is iterating on that queue.

Also removed "this will be fixed in future releases," because PCF legal dept. advises against making such promises in our docs.